ncclResult_t return error a string for better understanding of the error from the logs
Summary: This diff is part of a larger effort to improve the observability of errors in the NCCL library. Specifically, the code changes in the "checks.h" file modify the way errors are handled and reported. The changes include adding a call to "ncclGetErrorString" to convert the error code into a human-readable string, which is then included in the error message printed to the console. This should make it easier for developers to diagnose and fix issues with the NCCL library.

Reviewed By: wesbland

Differential Revision: D54678883

fbshipit-source-id: 7e9d7ee6015aeb546da3d9704d50d2dd193245b2