ctran: fix illegal mem access in sendrecv kernrel
Summary:
devState_d was passed to sendrecv kernel and the kernel accesses it to check rank/localRank metadata. However, devState_d was not set in sendrecv algorithm by mistake, thus causing illegal mem access.

One thing to notice is that, this error was not captured by internal CI because CI is built with mode/opt, and `assert` call is omitted. In contrast, `assert` is still valid when built in OSS mode and hence, caused `illegal mem access` when using in xlformer.

This patch fixes the issue, and made two changes to prevent similar issues:
- Check `devState_d` has been set before submitting kernel
- Remove `assert` in kernel and rely on collective algorithm to check before submitting kernel

Reviewed By: steinbrecher

Differential Revision: D55287919

fbshipit-source-id: 846ad158aabe24da6d9a221a5ad40c05f990a61a