Zero copy all to all: sender/receiver sync issue when canCopy16 mismatch
Summary:
When the send and receive have a mismatch on the canCopy16, the groupIdx that owns a portion of the shared memory buffer can differ. This can lead to a race condition. Consider the case where there are just 4 thread blocks (2 for sends and 2 for receives) and there are 32 threads/block. Let's say that send thread block 1 is very slow.
1) Send block 0 writes its first portion to shared memory.
2) Send block 0 writes its second portion to shared memory.
3) Send block 0 is done and sets its flag for receive block 0 to start. Meanwhile, send thread block 1 still hasn't started.
4) Receive block 0 is operating on int64s instead of uint4s. This means that every loop pulls 1/2 of the data that the send side wrote. It also means that it strides the shared memory buffer differently. This first copy is okay okay because the send thread block 0 had already put data in this spot of the shared buffer.
5) This is an out of sync memory read. Send thread block 1 has not yet written to this spot in the shared buffer, but receive thread block 0 is reading from it anyway because of the inconsistent striding between the writes and reads to the shared buffer.
{F1469907569}

The fix is to have each groupIdx own a portion of the shared memory buffer so that we cannot have this race condition.
 {F1474071751}

Reviewed By: minsii

Differential Revision: D55000136

fbshipit-source-id: 1e100ba2a1723e5f7331ce1c29aa5ff2de578c4d