Add nccl parameter to override cts
Summary: This change overrides the traffic class attribute of the QP on which the cts/fifo messages are sent on. This change will only be enabled when NCCL_IB_FIFO_TC environment variable is set otherwise its disabled by default. Setting NCCL_IB_FIFO_TC=192  changes the dscp value of the cts packets to 48 which maps to the network control queue(queue 7)

Reviewed By: wesbland

Differential Revision: D55926818

fbshipit-source-id: deed0c8c7273bcf100c6d383821c5628326adadc